+++ src/force.cpp
@@ -1,14 +1,14 @@
 #include "force.h"
 
-force::force(unsigned long long pid1, unsigned long long pid2, Eigen::Vector3d pos1, Eigen::Vector3d pos2, Eigen::Vector3d val) {
+force::force(unsigned long long pid1, unsigned long long pid2, Eigen::Vector3f pos1, Eigen::Vector3f pos2, Eigen::Vector3f val) {
   _pid1 = pid1;
   _pid2 = pid2;
   _pos1 = pos1;
   _pos2 = pos2;
   _val = val;
-  _dr = Eigen::Vector3d::Zero();
-  _dz = Eigen::Vector3d::Zero();
-  _df = Eigen::Vector3d::Zero();
+  _dr = Eigen::Vector3f::Zero();
+  _dz = Eigen::Vector3f::Zero();
+  _df = Eigen::Vector3f::Zero();
   _bandR = -1; _bandZ=-1; _bandN=-1;
   _dist = -1; _height = -1;
   _disable = false;
@@ -17,13 +17,13 @@ force::force(unsigned long long pid1, unsigned long long pid2, Eigen::Vector3d p
 
 force::force() {
   _pid1 = -1; _pid2 = -1;
-  _pos1 = Eigen::Vector3d::Zero();
-  _pos2 = Eigen::Vector3d::Zero();
-  _val = Eigen::Vector3d::Zero();
-  _dr = Eigen::Vector3d::Zero();
-  _dz = Eigen::Vector3d::Zero();
-  _df = Eigen::Vector3d::Zero();
-  _cP = Eigen::Vector3d::Zero();
+  _pos1 = Eigen::Vector3f::Zero();
+  _pos2 = Eigen::Vector3f::Zero();
+  _val = Eigen::Vector3f::Zero();
+  _dr = Eigen::Vector3f::Zero();
+  _dz = Eigen::Vector3f::Zero();
+  _df = Eigen::Vector3f::Zero();
+  _cP = Eigen::Vector3f::Zero();
   _bandR = -1; _bandZ=-1; _bandN=-1;
   _dist = -1; _height = -1;
   _disable = false;
