+++ src/particle.cpp
@@ -1,16 +1,16 @@
 #include "particle.h"
 #include <iostream>
 
-particle::particle(unsigned long long id, int type, double rad, Eigen::Vector3d c, Eigen::Vector3d v, Eigen::Vector3d o) {
+particle::particle(unsigned long long id, int type, double rad, Eigen::Vector3f c, Eigen::Vector3f v, Eigen::Vector3f o) {
   _id = id;
   _type = type;
   _rad = rad;
   _c = c;
   _v = v;
   _o = o;
-  _dr = Eigen::Vector3d::Zero();
-  _dz = Eigen::Vector3d::Zero();
-  _df = Eigen::Vector3d::Zero();
+  _dr = Eigen::Vector3f::Zero();
+  _dz = Eigen::Vector3f::Zero();
+  _df = Eigen::Vector3f::Zero();
   _dist = -1; _height = -1;
   _disable = false;
 };
@@ -19,12 +19,12 @@ particle::particle() {
   _id = -1;
   _type = -1;
   _rad = -1;
-  _c = Eigen::Vector3d::Zero();
-  _v = Eigen::Vector3d::Zero();
-  _o = Eigen::Vector3d::Zero();
-  _dr = Eigen::Vector3d::Zero();
-  _dz = Eigen::Vector3d::Zero();
-  _df = Eigen::Vector3d::Zero();
+  _c = Eigen::Vector3f::Zero();
+  _v = Eigen::Vector3f::Zero();
+  _o = Eigen::Vector3f::Zero();
+  _dr = Eigen::Vector3f::Zero();
+  _dz = Eigen::Vector3f::Zero();
+  _df = Eigen::Vector3f::Zero();
   _dist = -1; _height = -1;
   _disable = false;
 };
